refactor(builder): let the inspector own the canvas read, not the style panel
The `Changed files` gate failed with one introduced finding: `StyleInspectorPanel`
at cognitive 16 against a threshold of 15. Measured rather than guessed at —
removing only the hook call took the whole audit to `pass`, and the same
function carries no finding at all on clean `origin/main`, so it was mine and it
was that line.

Two attempts before this one made it worse or did nothing, which is what
guessing at a metric looks like: folding the tag read and the subject derivation
into one hook left it at 16, and extracting the form-choice state into a second
hook took it to 17.

The shape that fixes it is also the better one. Reading the rendered tag needs a
subscription to the canvas, and the panel that decides which control shows a
value has no business holding one. `InspectorPanel` owns the hook and passes the
ANSWER down — the same arrangement `cascade`, `breakpoints` and
`liveBreakpoints` already have, where the wrapper resolves and the panel
consumes. So the style panel takes `renderedTag?: string` instead of a canvas
ref, and holds no effect.

`useRenderedTag` moves to its own module for the same reason it moved out of
that component, and keeps the two things that are easy to get wrong: it reads
after a COMMIT, because the canvas and the inspector re-render together and the
DOM still holds the previous tag while a render body runs; and the document is a
dependency, because the tag is a function of the node's props.

Audit now reports zero introduced across all four categories, not just
complexity — the category a fix lands in is not always the one that failed.
